+++ embed_xray_other.go
@@ -1,7 +1,0 @@
-//go:build !windows
-
-package main
-
-var xrayEXE []byte
-var geoipDAT []byte
-var geositeDAT []byte